refactor 진척률 퍼사드 구조 수정

+++ src/main/java/com/ideality/coreflow/project/command/application/service/TaskService.java
@@ -23,7 +23,7 @@ public interface TaskService {
 
     void validateTask(Long taskId);
 
-    Double updateTaskProgress(Long taskId);
+    Long updateTaskProgress(Long taskId);
 
     String findTaskNameById(long taskId);
